+++ packages/demand-api/src/fare/FareService.ts
@@ -1,7 +1,7 @@
 import { Http } from '../http/types'
-import { FinalFareResponse } from './types'
+import { FinalFareResponse, Fare } from './types'
 
-export class FareService {
+export class FareService implements Fare {
   private url = 'fares'
 
   private http: Http