+++ packages/demand-api/src/trip/TripService.ts
@@ -1,5 +1,6 @@
 import { Http } from '../http/types'
 import {
+  Trip,
   TripFollowResponse,
   BookATripParams,
   BookATripWithoutNonceParams,
@@ -11,7 +12,7 @@ import {
   SearchResponse,
 } from './types'
 
-export class TripService {
+export class TripService implements Trip {
   private url = 'bookings'
 
   private http: Http