+++ packages/demand-bloc-track-trip/src/TrackTripBloc.test.ts
@@ -1,0 +1,7 @@
+import { TrackTripBloc } from './TrackTripBloc'
+
+describe('TrackTripBloc', () => {
+  it('should be a function', () => {
+    expect(typeof TrackTripBloc).toBe('function')
+  })
+})